[Mod] 파일구조변경
