+++ todo/package.json
@@ -13,8 +13,10 @@
     "@tanstack/react-query": "^5.60.5",
     "@tanstack/react-query-devtools": "^5.60.5",
     "axios": "^1.7.7",
+    "lodash": "^4.17.21",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.28.0",
     "styled-components": "^6.1.13"
   },
   "devDependencies": {
